Add basic plotter view

+++ discretefunction.cpp
@@ -5,14 +5,14 @@ DiscreteFunction::DiscreteFunction(QObject *parent)
 {
 }
 
-QMap<int, int> *DiscreteFunction::getDots()
+QMap<double, double> *DiscreteFunction::getDots()
 {
     return _dots;
 }
 
 void DiscreteFunction::arrayOfDots(BaseOperation *baseOp)
 {
-    _dots = new QMap<int, int>();
+    _dots = new QMap<double, double>();
     _baseOp = baseOp;
     for (double i = 0; i < 100; i++)
     {
